cosim(build): add '-Werror=conversion' and fix/cast surfaced issues; update isa sim submodule

+++ cosim/cosim.cpp
@@ -82,7 +82,7 @@ void cosim_force_irq(char mtip, char meip) {
 }
 
 DPI_LINKER_DECL DPI_DLLESPEC
-uint32_t cosim_get_inst_cnt() {
+uint64_t cosim_get_inst_cnt() {
     return rv32->get_inst_cnt();
 }
